Should work on non-centos linuxes now

+++ build.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env bash
 set -e
 
-# Copy local root certificates for corporate networks
-cp -u /etc/pki/ca-trust/source/anchors/* cert/
+# Copy local CentOS root certificates for wget to work behind corporate firewalls
+[ -e /etc/pki/ca-trust/source/anchors ] && cp -u /etc/pki/ca-trust/source/anchors/* cert/
 
 # Build base: base centos packages and gcc
 docker build -t "usd-docker/base:7.1" -f centos-6/base/Dockerfile .
